chore(all): clean-up (#30)
* Remove /// <reference />
This is not required as TS is now smart enough to include all .d.ts in your source folder, from used modules in node_modules and in @types org in node_modules.

* Fix aurelia-tools for Yarn
github:aurelia/tools#cli doesn't work (yet?) in yarn, but https://github.com/aurelia/tools.git#cli does.

* Fix "can be undefined" compile error
TS thinks that `plugin.hot` might have changed between the check and when the callback is called. We know it's not the case so I added a not-null assertion to remove the compilation error.

* Extracted webpack runtime types in their own file.
At the same time, I tweaked some definitions:
- `e` returns a Promise for nothing;
- `e` changed the function declaration style, which actually changes the expectations for `this` type. It implies the functions _have to be called on the `__webpack_require__` object_.

* Removed useless .filter()
`getOwnPropertyNames()` is always an array of string.
If you think of numbers (array?) their are converted to strings. If you think of Symbols, they use a distinct API.

* Removed useless try.catch
An exception thrown inside a Promise results in its rejection, no need to explicitly do it.

* Removed function allocation
As there was no code left...