chore(build): do not bundle HMR when not needed
I have cut the static dependency on `aurelia-hot-module-reload` and moved it inside the `if (module.hot)` branch.
This is detected by Webpack, so that `aurelia-hot-module-reload` is only bundled when compiled with `--hot`.
This helps reduce the size of production builds.

One thing that I'm not happy with is the literal typing of field `hmrContext`. But to improve that I think we need to export `HmrContext` definition in the DTS of `aurelia-hot-module-reload` in a way that can be consumed from the DTS only with no import (e.g. define in a namespace rather than a module?).